#465 rename to fcgi.h
